Correct Response format implemented & function foundation implemented

+++ Codebase/DatabaseBackend/backend/src/main/java/com/backend/terminplanungsassitent/exceptions/TerminNotFoundException.java
@@ -2,7 +2,7 @@ package com.backend.terminplanungsassitent.exceptions;
 
 public class TerminNotFoundException extends RuntimeException {
 
-    public TerminNotFoundException(Integer id) {
+    public TerminNotFoundException(Long id) {
         super("Could not find calendar " + id);
     }
